SearchExEditor component create
